Update README.md
At Solution Informatique, we are proud of our highly qualified technical team, consisting of engineers, supervisors, and experienced technicians. Alongside our sales, support, and after-sales team, we have the experience and resources to design and implement security systems that meet our clients' unique needs. With years of experience in system security, we have built a solid reputation for providing professional and effective solutions to ensure our clients' security and protection.